Added EmployeeEdit that currently just outputs updated data to console, in next lesson, we will add logic to update record in Firegase.
